Fix format string mismatch

+++ Code/Tools/FBuild/FBuildCore/Graph/ObjectNode.cpp
@@ -1387,7 +1387,7 @@ void ObjectNode::WriteToCache( Job * job )
                     AStackString<> output;
                     output.Format( "Obj: %s\n"
                                    " - Cache Store: %u ms (Store: %u ms - Compress: %u ms) (Compressed: %zu - Uncompressed: %zu) '%s'\n",
-                                   GetName().Get(), cachingTime, ( stopPublish - startPublish ), ( stopCompress - startCompress ), dataSize, buffer.GetDataSize(), cacheFileName.Get() );
+                                   GetName().Get(), cachingTime, ( stopPublish - startPublish ), ( stopCompress - startCompress ), dataSize, (size_t)buffer.GetDataSize(), cacheFileName.Get() );
                     if ( m_PCHCacheKey != 0 )
                     {
                         output.AppendFormat( " - PCH Key: %" PRIx64 "\n", m_PCHCacheKey );
